[~] Make ThreadHandle an intptr_t, so it's safe for 64-bit builds
git-svn-id: https://subversion.assembla.com/svn/Daedalusx64/trunk@1518 42e9bfbe-799a-4a2d-bad1-236e862a387a

+++ Source/Utility/Thread.h
@@ -27,7 +27,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 typedef u32 ( DAEDALUS_THREAD_CALL_TYPE * DaedThread )( void * arg );
 
-typedef s32 ThreadHandle;
+typedef intptr_t ThreadHandle;
 extern const ThreadHandle	kInvalidThreadHandle;
 
 enum EThreadPriority
